Client : Refactor HintText.vue. Repositioned
