+++ client/src/components/views/seb-server/settings-navigation/assessment-tool/create-assessment-tool/CreateAssessmentTool.vue
@@ -144,7 +144,6 @@ import { ref } from "vue";
 import BasicSettingsPage from "@/components/layout/pages/BasicSettingsPage.vue";
 import FormBuilder from "@/components/widgets/formBuilder/FormBuilder.vue";
 import LoadingFallbackComponent from "@/components/widgets/loadingFallbackComponent/LoadingFallbackComponent.vue";
-import HintText from "@/components/views/seb-server/settings-navigation/widgets/HintText.vue";
 import { navigateToRoute } from "@/router/navigation";
 import { useAssessmentToolFormFields } from "./composable/useAssessmentToolFormFields";
 import { useMutation } from "@/composables/useMutation";
@@ -152,6 +151,7 @@ import { createAssessmentTool } from "@/services/seb-server/assessmentToolServic
 import type { CommonAssessmentToolPar } from "@/models/seb-server/assessmentTool";
 import CancelButton from "@/components/widgets/CancelButton.vue";
 import ConfirmButton from "@/components/widgets/ConfirmButton.vue";
+import HintText from "@/components/widgets/HintText.vue";
 
 const {
     mainFormFields,
