+++ client/src/components/views/seb-server/settings-navigation/user-account/create-user-account/CreateUserAccount.vue
@@ -77,7 +77,6 @@ import { computed, ref } from "vue";
 import BasicSettingsPage from "@/components/layout/pages/BasicSettingsPage.vue";
 import FormBuilder from "@/components/widgets/formBuilder/FormBuilder.vue";
 import LoadingFallbackComponent from "@/components/widgets/loadingFallbackComponent/LoadingFallbackComponent.vue";
-import HintText from "@/components/views/seb-server/settings-navigation/widgets/HintText.vue";
 import UserAccountPreviewCard from "@/components/views/seb-server/settings-navigation/widgets/UserAccountPreviewCard.vue";
 import { useUserAccountFormFields } from "./composable/useUserAccountFormFields";
 import { useMutation } from "@/composables/useMutation";
@@ -87,6 +86,7 @@ import { UserRoleEnum } from "@/models/userRoleEnum";
 import i18n from "@/i18n";
 import CancelButton from "@/components/widgets/CancelButton.vue";
 import ConfirmButton from "@/components/widgets/ConfirmButton.vue";
+import HintText from "@/components/widgets/HintText.vue";
 
 const {
     leftFormFields,
