+++ client/src/components/widgets/HintText.vue
